requirements.txt
feedparser
requests
google-generativeai
python-dotenv
beautifulsoup4
dateparser
